sim: bfin: cast pointers using uintptr_t
We can't assume that sizeof(long) == sizeof(void*), so change all
these casts over to uintptr_t.

+++ sim/bfin/dv-bfin_cec.c
@@ -131,7 +131,7 @@ bfin_cec_io_read_buffer (struct hw *me, void *dest,
     return 0;
 
   mmr_off = addr - cec->base;
-  valuep = (void *)((unsigned long)cec + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)cec + mmr_base() + mmr_off);
 
   HW_TRACE_READ ();
