+++ sim/bfin/dv-bfin_eppi.c
@@ -101,7 +101,7 @@ bfin_eppi_io_write_buffer (struct hw *me, const void *source,
     value = dv_load_2 (source);
 
   mmr_off = addr - eppi->base;
-  valuep = (void *)((unsigned long)eppi + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)eppi + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
 
@@ -162,7 +162,7 @@ bfin_eppi_io_read_buffer (struct hw *me, void *dest,
     return 0;
 
   mmr_off = addr - eppi->base;
-  valuep = (void *)((unsigned long)eppi + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)eppi + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
