+++ sim/bfin/dv-bfin_mmu.c
@@ -109,7 +109,7 @@ bfin_mmu_io_write_buffer (struct hw *me, const void *source,
   value = dv_load_4 (source);
 
   mmr_off = addr - mmu->base;
-  valuep = (void *)((unsigned long)mmu + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)mmu + mmr_base() + mmr_off);
 
   HW_TRACE_WRITE ();
 
@@ -183,7 +183,7 @@ bfin_mmu_io_read_buffer (struct hw *me, void *dest,
     return 0;
 
   mmr_off = addr - mmu->base;
-  valuep = (void *)((unsigned long)mmu + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)mmu + mmr_base() + mmr_off);
 
   HW_TRACE_READ ();
