+++ sim/bfin/dv-bfin_nfc.c
@@ -84,7 +84,7 @@ bfin_nfc_io_write_buffer (struct hw *me, const void *source, int space,
 
   value = dv_load_2 (source);
   mmr_off = addr - nfc->base;
-  valuep = (void *)((unsigned long)nfc + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)nfc + mmr_base() + mmr_off);
 
   HW_TRACE_WRITE ();
 
@@ -134,7 +134,7 @@ bfin_nfc_io_read_buffer (struct hw *me, void *dest, int space,
     return 0;
 
   mmr_off = addr - nfc->base;
-  valuep = (void *)((unsigned long)nfc + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)nfc + mmr_base() + mmr_off);
 
   HW_TRACE_READ ();
