+++ sim/bfin/dv-bfin_ppi.c
@@ -92,7 +92,7 @@ bfin_ppi_io_write_buffer (struct hw *me, const void *source, int space,
 
   value = dv_load_2 (source);
   mmr_off = addr - ppi->base;
-  valuep = (void *)((unsigned long)ppi + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)ppi + mmr_base() + mmr_off);
 
   HW_TRACE_WRITE ();
 
@@ -131,7 +131,7 @@ bfin_ppi_io_read_buffer (struct hw *me, void *dest, int space,
     return 0;
 
   mmr_off = addr - ppi->base;
-  valuep = (void *)((unsigned long)ppi + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)ppi + mmr_base() + mmr_off);
 
   HW_TRACE_READ ();
