+++ sim/bfin/dv-bfin_sic.c
@@ -161,7 +161,7 @@ bfin_sic_52x_io_write_buffer (struct hw *me, const void *source, int space,
     value = dv_load_2 (source);
 
   mmr_off = addr - sic->base;
-  valuep = (void *)((unsigned long)sic + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)sic + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
 
@@ -214,7 +214,7 @@ bfin_sic_52x_io_read_buffer (struct hw *me, void *dest, int space,
     return 0;
 
   mmr_off = addr - sic->base;
-  valuep = (void *)((unsigned long)sic + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)sic + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
 
@@ -275,7 +275,7 @@ bfin_sic_537_io_write_buffer (struct hw *me, const void *source, int space,
     value = dv_load_2 (source);
 
   mmr_off = addr - sic->base;
-  valuep = (void *)((unsigned long)sic + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)sic + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
 
@@ -327,7 +327,7 @@ bfin_sic_537_io_read_buffer (struct hw *me, void *dest, int space,
     return 0;
 
   mmr_off = addr - sic->base;
-  valuep = (void *)((unsigned long)sic + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)sic + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
 
@@ -389,7 +389,7 @@ bfin_sic_54x_io_write_buffer (struct hw *me, const void *source, int space,
     value = dv_load_2 (source);
 
   mmr_off = addr - sic->base;
-  valuep = (void *)((unsigned long)sic + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)sic + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
 
@@ -438,7 +438,7 @@ bfin_sic_54x_io_read_buffer (struct hw *me, void *dest, int space,
     return 0;
 
   mmr_off = addr - sic->base;
-  valuep = (void *)((unsigned long)sic + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)sic + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
 
@@ -496,7 +496,7 @@ bfin_sic_561_io_write_buffer (struct hw *me, const void *source, int space,
     value = dv_load_2 (source);
 
   mmr_off = addr - sic->base;
-  valuep = (void *)((unsigned long)sic + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)sic + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
 
@@ -549,7 +549,7 @@ bfin_sic_561_io_read_buffer (struct hw *me, void *dest, int space,
     return 0;
 
   mmr_off = addr - sic->base;
-  valuep = (void *)((unsigned long)sic + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)sic + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
