+++ sim/bfin/dv-bfin_uart2.c
@@ -83,7 +83,7 @@ bfin_uart_io_write_buffer (struct hw *me, const void *source,
 
   value = dv_load_2 (source);
   mmr_off = addr - uart->base;
-  valuep = (void *)((unsigned long)uart + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)uart + mmr_base() + mmr_off);
 
   HW_TRACE_WRITE ();
 
@@ -140,7 +140,7 @@ bfin_uart_io_read_buffer (struct hw *me, void *dest,
     return 0;
 
   mmr_off = addr - uart->base;
-  valuep = (void *)((unsigned long)uart + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)uart + mmr_base() + mmr_off);
 
   HW_TRACE_READ ();
