+++ sim/bfin/dv-bfin_wdog.c
@@ -72,7 +72,7 @@ bfin_wdog_io_write_buffer (struct hw *me, const void *source,
     value = dv_load_2 (source);
 
   mmr_off = addr - wdog->base;
-  valuep = (void *)((unsigned long)wdog + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)wdog + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
 
@@ -121,7 +121,7 @@ bfin_wdog_io_read_buffer (struct hw *me, void *dest,
     return 0;
 
   mmr_off = addr - wdog->base;
-  valuep = (void *)((unsigned long)wdog + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)wdog + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
