+++ sim/bfin/dv-bfin_wp.c
@@ -84,7 +84,7 @@ bfin_wp_io_write_buffer (struct hw *me, const void *source, int space,
 
   value = dv_load_4 (source);
   mmr_off = addr - wp->base;
-  valuep = (void *)((unsigned long)wp + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)wp + mmr_base() + mmr_off);
 
   HW_TRACE_WRITE ();
 
@@ -124,7 +124,7 @@ bfin_wp_io_read_buffer (struct hw *me, void *dest, int space,
     return 0;
 
   mmr_off = addr - wp->base;
-  valuep = (void *)((unsigned long)wp + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)wp + mmr_base() + mmr_off);
 
   HW_TRACE_READ ();
