+++ sim/bfin/dv-eth_phy.c
@@ -86,7 +86,7 @@ eth_phy_io_write_buffer (struct hw *me, const void *source,
   value = dv_load_2 (source);
 
   reg_off = addr - phy->base;
-  valuep = (void *)((unsigned long)phy + reg_base() + reg_off);
+  valuep = (void *)((uintptr_t)phy + reg_base() + reg_off);
 
   HW_TRACE_WRITE ();
 
@@ -117,7 +117,7 @@ eth_phy_io_read_buffer (struct hw *me, void *dest,
   bu16 *valuep;
 
   reg_off = addr - phy->base;
-  valuep = (void *)((unsigned long)phy + reg_base() + reg_off);
+  valuep = (void *)((uintptr_t)phy + reg_base() + reg_off);
 
   HW_TRACE_READ ();
